TST: Begin rewrite of tests
